- Added support for Raidable Boats by VisEntities - Changed amount, amountMin, and stacksize types to int from long

+++ RustRBLootEditor/Converters/FormatNumberConverter.cs
@@ -17,7 +17,7 @@ namespace RustRBLootEditor.Converters
                 return DependencyProperty.UnsetValue;
             else
             {
-                return FormatNumber((long)value);
+                return FormatNumber((int)value);
             }
         }
 
@@ -26,10 +26,10 @@ namespace RustRBLootEditor.Converters
             throw new NotImplementedException();
         }
 
-        private static string FormatNumber(long num)
+        private static string FormatNumber(int num)
         {
             // Ensure number has max 3 significant digits (no rounding up can happen)
-            long i = (long)Math.Pow(10, (int)Math.Max(0, Math.Log10(num) - 2));
+            int i = (int)Math.Pow(10, (int)Math.Max(0, Math.Log10(num) - 2));
             num = num / i * i;
 
             if (num >= 1000000000)
